pyproject.toml
[project]
name = "arras-contract-system"
version = "0.1.0"
description = "Controlled legal document engine for Catalonia arras contracts."
readme = "readme.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "rich>=13.7.0",
  "sqlalchemy>=2.0.0",
  "asyncpg>=0.29.0",
  "psycopg[binary]>=3.2.0",
  "alembic>=1.13.0",
  "jinja2>=3.1.0",
  "python-docx>=1.1.0",
  "weasyprint>=62.0",
  "openai>=1.40.0",
  "python-dotenv>=1.0.0",
  "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "httpx>=0.27.0",
  "ruff>=0.6.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = "-q"

[tool.uv]
package = false
